requirements.txt
Django==1.6.5
Pillow==2.4.0
South==0.8.4
argparse==1.2.1
django-filer==0.9.5
django-mptt==0.5.5
django-polymorphic==0.5.5
easy-thumbnails==2.0.1
wsgiref==0.1.2
